requirements.txt
langchain==0.1.0
langchain_community==0.0.16
langchain_core==0.1.17
langchain_openai==0.0.5
python-dotenv==1.0.1
streamlit==1.30.0
chromadb==0.4.22
beautifulsoup4=4.12.2
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
unstructured
# uncomment to use huggingface llms
huggingface-hub==0.14.1
python-magic-bin==0.4.14

# uncomment to use instructor embeddings

sentence-transformers==2.2.2